requirements.txt
python-telegram-bot[all]==21.5
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
